+++ jreunion/src/com/googlecode/reunion/jreunion/game/HumanPlayer.java
@@ -14,24 +14,24 @@ public class HumanPlayer extends Player {
 		super(client);
 	}
 
-	public int getMaxElectricity(){
+	public long getMaxElectricity(){
 		return Tools.statCalc(getDexterity(), 30) +(getLeadership() / 2);
 	}
 	
 	@Override
-	public int getMaxHp(){
+	public long getMaxHp(){
 		return Tools.statCalc(getStrength(), 80) + Tools.statCalc(getConstitution(), 30)+ (getLeadership() / 2);		
 	}
 	
-	public int getMaxMana(){
+	public long getMaxMana(){
 		return Tools.statCalc(getWisdom(), 50) + (getLeadership() / 2);		
 	}
-	public int getMaxStamina(){
+	public long getMaxStamina(){
 		return getStrength() + (getLeadership() / 2);
 	}
 
 	@Override
-	public float getBaseDamage() {
+	public long getBaseDamage() {
 		return (getLevel() / 6) + (getDexterity() / 4)+ getStrength();
 		
 	}
