+++ jreunion/src/com/googlecode/reunion/jreunion/game/skills/aidia/SafetyShield.java
@@ -99,8 +99,8 @@ public class SafetyShield extends Skill implements Castable, Effectable{
 		return 50f/(getMaxLevel()-1);
 	}
 	
-	float getManaModifier(Player player){
-		float modifier = 0;
+	long getManaModifier(Player player){
+		long modifier = 0;
 		int level = player.getSkillLevel(this);
 		
 		if(level>0){
@@ -121,8 +121,8 @@ public class SafetyShield extends Skill implements Castable, Effectable{
 		return 24f/(getMaxLevel()-1);
 	}
 	
-	float getDurationModifier(Player player){
-		float modifier = 0;
+	long getDurationModifier(Player player){
+		long modifier = 0;
 		int level = player.getSkillLevel(this);
 		
 		if(level>0){
@@ -143,8 +143,8 @@ public class SafetyShield extends Skill implements Castable, Effectable{
 		return 240f/(getMaxLevel()-1);
 	}
 	
-	float getAccumulatedTimeModifier(Player player){
-		float modifier = 0;
+	long getAccumulatedTimeModifier(Player player){
+		long modifier = 0;
 		int level = player.getSkillLevel(this);
 		
 		if(level>0){
@@ -161,9 +161,9 @@ public class SafetyShield extends Skill implements Castable, Effectable{
 		
 		if(caster instanceof AidiaPlayer){
 			final Player player = (Player) caster;
-			int newMana = player.getMana() - (int) getManaModifier(player);
+			long newMana = player.getMana() - getManaModifier(player);
 			
-			if(getEffectModifier() == (int)getAccumulatedTimeModifier(player)){
+			if(getEffectModifier() == getAccumulatedTimeModifier(player)){
 				
 				player.getClient().sendPacket(Type.SAY, "SafetyShield skill acumulated time, already at maximum.");
 				return false;
