+++ jreunion/src/com/googlecode/reunion/jreunion/game/skills/bulkan/WhirlwindSlash.java
@@ -66,8 +66,8 @@ public class WhirlwindSlash extends WeaponAttack implements Castable{
 		return 23f/(getMaxLevel()-1);
 	}
 	
-	float getStaminaModifier(Player player){
-		float modifier = 0;
+	long getStaminaModifier(Player player){
+		long modifier = 0;
 		int level = player.getSkillLevel(this);
 		
 		if(level>0){
@@ -82,14 +82,14 @@ public class WhirlwindSlash extends WeaponAttack implements Castable{
 		
 		if(caster instanceof BulkanPlayer){	
 			Player player = (Player)caster;
-			int currentStamina = player.getStamina();
-			int staminaSpent = (int) getStaminaModifier(player); 
+			long currentStamina = player.getStamina();
+			long staminaSpent = getStaminaModifier(player); 
 					
 			player.setStamina(currentStamina - staminaSpent);
 			
-			float baseDamage = player.getBaseDamage();
+			long baseDamage = player.getBaseDamage();
 			float skillDamage = getDamageModifier(player);
-			float weaponDamage = 0;
+			long weaponDamage = 0;
 			Item<?> weapon = player.getEquipment().getMainHand();
 			
 			if( weapon != null)
@@ -97,7 +97,7 @@ public class WhirlwindSlash extends WeaponAttack implements Castable{
 						(Server.getRand().nextFloat()*(((Weapon)weapon.getType()).getMaxDamage(weapon)
 								- ((Weapon)weapon.getType()).getMinDamage(weapon)));
 			
-			float damage = (baseDamage +  weaponDamage) * skillDamage;
+			long damage = (long)((baseDamage +  weaponDamage) * skillDamage);
 			
 			for(Skill skill: player.getSkills().keySet()){
 				if (Modifier.class.isInstance(skill)){
@@ -124,7 +124,7 @@ public class WhirlwindSlash extends WeaponAttack implements Castable{
 				
 			synchronized(victims){	
 				for(LivingObject victim : victims){ 
-					victim.getsAttacked(player, (int)damage);
+					victim.getsAttacked(player, damage);
 				}
 				return true;
 			}
